[Add/Update] startscenecontroller 스크립트 내용수정, 난이도 조절 버튼 삭제 및 패널로 대체 (-장범)
-StartSceneController 스크립트 수정
-난이도 조절 버튼 삭제 및 패널로 대체
-visual studio 빨간줄 안나오는 에러 (package 추가로 수정)
