+++ Assets/Scripts/StartSceneController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3e2f8f528b19534488b017fb46c620a2
+guid: 27b18b79a0ebf9045b0fa58ceb66ac74
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
