+++ Packages/manifest.json
@@ -1,6 +1,7 @@
 {
   "dependencies": {
     "com.unity.feature.2d": "2.0.0",
+    "com.unity.ide.visualstudio": "2.0.20",
     "com.unity.textmeshpro": "3.0.6",
     "com.unity.modules.ai": "1.0.0",
     "com.unity.modules.androidjni": "1.0.0",
